+++ lib/llvm/transforms/utils.rb
@@ -10,25 +10,21 @@ module LLVM
     # /** See llvm::createLowerSwitchPass function. */
     # void LLVMAddLowerSwitchPass(LLVMPassManagerRef PM);
     def lowerswitch!
-      C.add_lower_switch_pass(self)
+      raise DeprecationError
     end
 
     # @LLVMpass mem2reg
     # /** See llvm::createPromoteMemoryToRegisterPass function. */
     # void LLVMAddPromoteMemoryToRegisterPass(LLVMPassManagerRef PM);
     def mem2reg!
-      C.add_promote_memory_to_register_pass(self)
+      raise DeprecationError
     end
 
     # @LLVMpass add-discriminators
     # /** See llvm::createAddDiscriminatorsPass function. */
     # void LLVMAddAddDiscriminatorsPass(LLVMPassManagerRef PM);
     def add_discriminators!
-      C.add_add_discriminators_pass(self)
+      raise DeprecationError
     end
-  end
-
-  module C
-    attach_function :add_add_discriminators_pass, :LLVMAddAddDiscriminatorsPass, [:pointer], :void
   end
 end