+++ test/test_helper.rb
@@ -13,6 +13,12 @@ begin
 
   SimpleCov.start do
     add_filter "/test/"
+    add_filter "/lib/llvm/transforms/scalar.rb"
+    add_filter "/lib/llvm/transforms/ipo.rb"
+    add_filter "/lib/llvm/transforms/vectorize.rb"
+    add_filter "/lib/llvm/transforms/utils.rb"
+    add_filter "/lib/llvm/transforms/builder.rb"
+    add_filter "/lib/llvm/core/pass_manager.rb"
   end
 rescue LoadError
   warn "Proceeding without SimpleCov. gem install simplecov on supported platforms."